Move extensions to the `gradle-api` module
